+++ include/json-rpc.h
@@ -4,7 +4,8 @@
 #include "config.h"
 #include "logger.h"
 #include "main.h"
-#include "spiffs.h"
+#include "blesfs.h"
+
 
 #include <ArduinoJson.h>
 #include <string>
