+++ include/logger.h
@@ -2,7 +2,8 @@
 #define BLESKOMAT_LOGGER_H
 
 #include "config.h"
-#include "spiffs.h"
+#include "FS.h"
+#include "SPIFFS.h"
 #include <Arduino.h>
 #include <iostream>
 #include <map>
